[Chore] 폴더명 변경
