+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/auth/dto/request/RefreshTokenRequest.java
@@ -1,0 +1,8 @@
+package com.catchsolmind.cheongyeonbe.domain.auth.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RefreshTokenRequest(
+        @NotBlank String refreshToken
+) {
+}
