+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/auth/dto/response/RefreshTokenResponse.java
@@ -1,0 +1,12 @@
+package com.catchsolmind.cheongyeonbe.domain.auth.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record RefreshTokenResponse(
+        String accessToken,
+        Integer expiresIn,
+        String refreshToken,
+        Integer refreshTokenExpiresIn
+) {
+}
