+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/auth/service/AuthTokenService.java
@@ -1,0 +1,8 @@
+package com.catchsolmind.cheongyeonbe.domain.auth.service;
+
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.request.RefreshTokenRequest;
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.response.RefreshTokenResponse;
+
+public interface AuthTokenService {
+    RefreshTokenResponse refresh(RefreshTokenRequest request);
+}
