+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/auth/service/KakaoClientService.java
@@ -1,0 +1,13 @@
+package com.catchsolmind.cheongyeonbe.domain.auth.service;
+
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.response.KakaoTokenResponse;
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.response.KakaoUserResponse;
+
+public interface KakaoClientService {
+
+    KakaoTokenResponse requestToken(String code, String redirectUri);
+
+    String getDefaultRedirectUri();
+
+    KakaoUserResponse getKakaoUserInfo(String accessToken);
+}
