--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/dto/request/RefreshTokenRequest.java
@@ -1,8 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RefreshTokenRequest(
-        @NotBlank String refreshToken
-) {
-}
