--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/dto/response/KakaoLoginResponse.java
@@ -1,29 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.dto.response;
-
-import com.catchsolmind.cheongyeonbe.domain.user.dto.UserDto;
-import lombok.Builder;
-
-@Builder
-public record KakaoLoginResponse(
-        String accessToken,
-        Integer expiresIn,
-        String refreshToken,
-        Integer refreshTokenExpiresIn,
-        UserDto user
-) {
-    public static KakaoLoginResponse of(
-            UserDto userDto,
-            String accessToken,
-            long accessTokenExpirationMs,
-            String refreshToken,
-            long refreshTokenExpirationMs
-    ) {
-        return new KakaoLoginResponse(
-                accessToken,
-                (int) (accessTokenExpirationMs / 1000),
-                refreshToken,
-                (int) (refreshTokenExpirationMs / 1000),
-                userDto
-        );
-    }
-}
