--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/dto/response/RefreshTokenResponse.java
@@ -1,12 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record RefreshTokenResponse(
-        String accessToken,
-        Integer expiresIn,
-        String refreshToken,
-        Integer refreshTokenExpiresIn
-) {
-}
