--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/entity/RefreshToken.java
@@ -1,48 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.entity;
-
-import com.catchsolmind.cheongyeonbe.domain.user.entity.User;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "refresh_token",
-        indexes = {
-                @Index(name = "idx_refresh_token", columnList = "refresh_token")
-        }
-)
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "refresh_token_id")
-    private Long refreshTokenId;
-
-    @Column(name = "refresh_token", nullable = false, length = 500, unique = true)
-    private String refreshToken;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @Column(name = "expires_at", nullable = false)
-    private LocalDateTime expiresAt;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    // 편의 메서드,
-    public boolean isExpired() {
-        return !expiresAt.isAfter(LocalDateTime.now());
-    }
-}
