--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/service/AuthTokenService.java
@@ -1,8 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.service;
-
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.request.RefreshTokenRequest;
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.response.RefreshTokenResponse;
-
-public interface AuthTokenService {
-    RefreshTokenResponse refresh(RefreshTokenRequest request);
-}
