--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/service/KakaoAuthService.java
@@ -1,6 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.service;
-
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.response.KakaoLoginResponse;
-
-public interface KakaoAuthService {
-    KakaoLoginResponse login(String code, String redirectUri);}
