--- src/main/java/com/catchsolmind/cheongyeonbe/domain/oauth/service/KakaoClientService.java
@@ -1,13 +1,0 @@
-package com.catchsolmind.cheongyeonbe.domain.oauth.service;
-
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.response.KakaoTokenResponse;
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.response.KakaoUserResponse;
-
-public interface KakaoClientService {
-
-    KakaoTokenResponse requestToken(String code, String redirectUri);
-
-    String getDefaultRedirectUri();
-
-    KakaoUserResponse getKakaoUserInfo(String accessToken);
-}
