+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/user/entity/User.java
@@ -1,7 +1,7 @@
 package com.catchsolmind.cheongyeonbe.domain.user.entity;
 
 import com.catchsolmind.cheongyeonbe.domain.group.entity.GroupMember;
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.data.OAuthUserInfo;
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.data.OAuthUserInfo;
 import com.catchsolmind.cheongyeonbe.domain.point.entity.PointTransaction;
 import com.catchsolmind.cheongyeonbe.global.enums.AuthProvider;
 import jakarta.persistence.*;
