+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/user/service/UserService.java
@@ -1,6 +1,6 @@
 package com.catchsolmind.cheongyeonbe.domain.user.service;
 
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.data.OAuthUserInfo;
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.data.OAuthUserInfo;
 import com.catchsolmind.cheongyeonbe.domain.user.dto.UserDto;
 import com.catchsolmind.cheongyeonbe.domain.user.entity.User;
