+++ src/test/java/com/catchsolmind/cheongyeonbe/global/fixture/dto/oauth/KakaoLoginResponseFixture.java
@@ -1,6 +1,6 @@
 package com.catchsolmind.cheongyeonbe.global.fixture.dto.oauth;
 
-import com.catchsolmind.cheongyeonbe.domain.oauth.dto.response.KakaoLoginResponse;
+import com.catchsolmind.cheongyeonbe.domain.auth.dto.response.KakaoLoginResponse;
 import com.catchsolmind.cheongyeonbe.global.fixture.dto.user.UserDtoFixture;
 
 public class KakaoLoginResponseFixture {
